Fix chain counts in the chain parser utility
Was using atoi to convert from string to in number of counts, but for large numbers
counts was printed in scientific notation which was failing in the atoi, so converted
to atof and all is well


Former-commit-id: 72f0cc8e7c87051b03494c726f1d64bdf146b5f0

+++ Trigger/TrigAnalysis/TrigInDetAnalysisUser/Analysis/src/chainparser.cxx
@@ -297,7 +297,7 @@ int main( int argc, char** argv ) {
       
     if ( expl.size() > expected_size ) { 
 
-      int counts = std::atoi(expl[expected_size].c_str());
+      int counts = std::atof(expl[expected_size].c_str());
 
       /// ignore chains with no entries ( could print out as problem chains if required)
       if ( counts >= userthreshold ) {
